Add OxfordDictionary::Deserialize
For V2 we are going to stop snake casing the response. We will still be
able to navigate the response objects with dot syntax (since it's just
an OpenStruct), but the keys will look like JSON (camelCase) instead.

If someone really wants to, they could patch this library and use a
different deserializer in the endpoint classes. Maybe one day this is
configurable but I don't really care at the moment because I'm not sure
there are compelling reasons to do so.