Add Entries endpoint for V2
Oxford Dictionaries is updating their API to a new version which
includes quite a few changes:
https://developer.oxforddictionaries.com/version2

I've left this `#entry_snake_case` method so that if someone needs more
time to upgrade they can use this to keep snake cased keys in their
response objects. There's a more detailed message on the commit that
adds this to `OxfordDictionary::Client`'s interface.