Add OxfordDictionary::Endpoints::Search
Oxford Dictionaries is updating their API to a new version which
includes quite a few changes:
https://developer.oxforddictionaries.com/version2

There have been updates to the search functionality which is documented
in the migration guide linked above.

As mentioned in the tests, this only unit tests the endpoint. If someone
would like to contribute, please feel free to update and run the specs
(similar to what was done for the Entries and Lemmas endpoints).

Entries: https://github.com/swcraig/oxford-dictionary/pull/8
Lemmas: https://github.com/swcraig/oxford-dictionary/pull/10