Add OxfordDictionary::Client#search
Oxford Dictionaries is updating their API to a new version which
includes quite a few changes:
https://developer.oxforddictionaries.com/version2

There have been updates to the search functionality which is documented
in the migration guide linked above.

I've tried to be smart and preserve the interface so that users can
upgrade to V2 without any immediate breaking changes (as long as the
V1 usage was what was documented in the README).

Check out `OxfordDictionary::Endpoints::Search#search` for how to use
the new interface which uses named parameters.