Add basic V2 documentation to the README
Oxford Dictionaries has discontinued their V1 API (it will be taken
offline on June 30, 2019). As part of the upgrade of this wrapper the
new interface uses named parameters.

Reference the following PRs for the new usage:
- Entries: https://github.com/swcraig/oxford-dictionary/pull/8
- Translations: https://github.com/swcraig/oxford-dictionary/pull/12
- Sentences: https://github.com/swcraig/oxford-dictionary/pull/13
- Thesaurus: https://github.com/swcraig/oxford-dictionary/pull/14
- Lemmas: https://github.com/swcraig/oxford-dictionary/pull/10
- Search: https://github.com/swcraig/oxford-dictionary/pull/15

All of the response objects are OpenStructs with camelCase keys. This is
a departure from V1 (which used snake_case keys in the response).