Remove plissken as a dependency
We no longer translate the API responses into snake_case keys. We keep
the keys in camelCase so that the object from our endpoint calls is as
close to the JSON response as possible. It was always sort of weird
because the parameters for the endpoint queries were in camelCase but
everything came back in snake_case.